1.0.1

+++ EdyExample/MyAppController.m
@@ -223,7 +223,7 @@
 
 
 /// ⑤ Felicaデータの受信
-- (BOOL)felicaReceive:(USBDevicePasori *)pasori data:(felica_ans_t *)ans numBytesRead:(UInt32)numBytesRead
+- (BOOL)felicaReceive:(USBDevicePasori *)pasori data:(felica_ans_t *)ans numBytesRead:(uint32_t)numBytesRead
 {
 	BOOL	handled = NO;
 
@@ -277,13 +277,13 @@
 #pragma mark -
 
 
-- (int)numberOfRowsInTableView:(NSTableView *)tableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)tableView
 {
 	return [edyValues count];
 }
 
 
-- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(int)row
+- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row
 {
 	return [[edyValues objectAtIndex:row] objectForKey:[tableColumn identifier]];
 }
